Restrict framework status updates
 ## Summary
A framework's status should only be able to move forwards over time. We
don't know any legitimate reason why (outside of our testing
environments), a framework status should ever need to go back a step. We
had an incident where a framework was briefly set back to the initial
state, and it would be good to add protections against this happening
again.