Use Flask built-in CLI instead of deprecated Flask-Script
Flask 1.x includes a command line tool for running Flask applictions
that replaces the deprecated Flask-Cli plugin.

This commit changes the Makefile to use this; the Flask-Cli
configuration is still present and can be used, but we should prefer to
use the new and maintained tool.

To keep the same preferences for port and ensure that the flask script
runs the right application, this commit also adds a `.flaskenv` file and
`python-dotenv` as a dev requirement; see the [Flask dotenv documentation]
for details on how this works.

[Flask dotenv documentation]: https://flask.palletsprojects.com/en/1.1.x/cli/#environment-variables-from-dotenv